fix: current lsp added in fullCourse
